wip refactor react code
