+++ react/src/components/context.js
@@ -1,9 +1,10 @@
 import React from 'react';
 // TODO: remove this file by making context usage implicit
-export const ContextReduxFormActions = React.createContext();
-export const ContextReduxFormSelectors = React.createContext();
+//export const ContextReduxFormActions = React.createContext();
+//export const ContextReduxFormSelectors = React.createContext();
 
 export const ContextModels = React.createContext();
+export const ContextRelativeFormState = React.createContext();
 
 export const ContextSimulationListPromise = React.createContext();
 export const ContextSimulationInfoPromise = React.createContext();
@@ -16,3 +17,8 @@ export const ContextRelativeRouterHelper = React.createContext();
 export const ContextRelativeFormController = React.createContext();
 export const ContextRelativeHookedDependencyGroup = React.createContext();
 export const ContextRelativeFormDependencies = React.createContext();
+
+//export const ContextRelativeViewName = React.createContext();
+//export const ContextRelativeViewConfig = React.createContext();
+
+export const ContextSchema = React.createContext();
