+++ react/src/components/graph2d.js
@@ -1,60 +1,3 @@
-// import { React }
-import { Zoom } from '@visx/zoom';
-import { Axis, ClipPath, Scale, Shape } from '@visx/visx';
-
-export function apiResponseToGraph2dConfig({
-    plots,
-    title,
-    x_label: xLabel,
-    x_points: xPoints,
-    x_range: xRange,
-    y_range: yRange,
-    y_label: yLabel
-}) {
-    let tempPlots = plots.map(({ color, label, points }) => {
-        let tempPoints = points.map((y, i) => { return { x: xPoints[i], y } })
-        return {
-            color,
-            label,
-            points: tempPoints
-        }
-    });
-
-    let [xMin, xMax] = xRange;
-    let [yMin, yMax] = yRange;
-
-    return {
-        title,
-        plots: tempPlots,
-        xLabel,
-        yLabel,
-        xRange: {
-            min: xMin,
-            max: xMax
-        },
-        yRange: {
-            min: yMin,
-            max: yMax
-        }
-    }
-}
-
-export let Graph2dFromApi = {
-    getDependencies: (config) => {
-        return [];
-    },
-
-    element: (props) => {
-        let { simulationData } = props;
-
-        let config = apiResponseToGraph2dConfig(simulationData);
-
-        return (
-            <Graph2d {...config} {...props}/>
-        )
-    }
-}
-
 /**
  *
  * @param {{
