+++ react/src/components/panel.js
@@ -1,25 +1,3 @@
-import { Card, Modal, Col, Button, Form, Tab, Tabs } from "react-bootstrap";
-import { useState, Fragment, useContext } from 'react';
-import { useStore } from "react-redux";
-import { FormController, FieldGridLayout, FieldListLayout } from "./form";
-import { Dependency, HookedDependencyGroup, useInterpolatedString } from "../dependency";
-import {
-    ContextRelativeHookedDependencyGroup,
-    ContextRelativeFormController,
-    ContextReduxFormActions,
-    ContextReduxFormSelectors,
-    ContextSimulationInfoPromise,
-    ContextModels as ContextRelativeModels,
-    ContextRelativeFormDependencies
-} from "./context";
-import * as Icon from "@fortawesome/free-solid-svg-icons";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { useRenderCount } from "../hooks";
-
-import "./panel.scss";
-import { AutoRunReportLayout } from "./report";
-import { Graph2dFromApi } from "./graph2d";
-
 export function Panel(props) {
     let { title, buttons, panelBodyShown, ...otherProps } = props;
     return (
@@ -49,14 +27,6 @@ function ViewPanelActionButtons(props) {
     )
 }
 
-function EditorForm(props) {
-    return (
-        <Form>
-            {props.children}
-        </Form>
-    );
-}
-
 export function EditorPanel(props) {
     let {
         submit,
@@ -68,9 +38,6 @@ export function EditorPanel(props) {
         title,
         id
     } = props;
-
-    useRenderCount("EditorPanel");
-
     let [advancedModalShown, updateAdvancedModalShown] = useState(false);
     let [panelBodyShown, updatePanelBodyShown] = useState(true);
 
@@ -120,187 +87,4 @@ export function EditorPanel(props) {
             {showButtons && actionButtons}
         </Panel>
     )
-}
-
-export let TabLayout = {
-    getDependencies: (config) => {
-        let fields = [];
-
-        for (let tab of config.tabs) {
-            for (let layoutConfig of tab.items) {
-                let ele = elementForLayoutName(layoutConfig.layout);
-                fields.push(...ele.getDependencies(layoutConfig));
-            }
-        }
-
-        return fields;
-    },
-
-    element: (props) => {
-        let { config } = props;
-
-        let renderCountFn = useRenderCount;
-        renderCountFn("TabLayout");
-
-        let tabs = config.tabs;
-
-        let tabEls = [];
-
-        let firstTabKey = undefined;
-
-        for (let tabConfig of tabs) {
-            let name = tabConfig.name;
-            let layouts = tabConfig.items;
-            let layoutElements = layouts.map((layoutConfig, idx) => {
-                let ele = elementForLayoutName(layoutConfig.layout)
-                let LayoutElement = ele.element;
-                return <LayoutElement key={idx} config={layoutConfig}></LayoutElement>
-            })
-            firstTabKey = firstTabKey || name;
-            tabEls.push(
-                <Tab key={name} eventKey={name} title={name}>
-                    {layoutElements}
-                </Tab>
-            )
-        }
-
-        return (
-            <Tabs defaultActiveKey={firstTabKey}>
-                {tabEls}
-            </Tabs>
-        )
-    }
-}
-
-function SpacedLayout(layoutElement) {
-    let ChildElement = layoutElement.element;
-    return {
-        getDependencies: layoutElement.getDependencies,
-
-        element: (props) => {
-            let renderCountFn = useRenderCount;
-            renderCountFn("SpacedLayout");
-
-            return (
-                <div className="sr-form-layout">
-                    <ChildElement {...props} />
-                </div>
-            )
-        }
-    }
-}
-
-export let MissingLayout = {
-    getDependencies: () => {
-        return [];
-    },
-
-    element: (props) => {
-        return <>Missing layout!</>;
-    }
-}
-
-let layoutElements = {
-    "fieldList": SpacedLayout(FieldListLayout),
-    "fieldTable": SpacedLayout(FieldGridLayout),
-    "tabs": TabLayout,
-    "graph2d": AutoRunReportLayout(Graph2dFromApi)
-}
-
-export function elementForLayoutName(layoutName) {
-    return layoutElements[layoutName] || MissingLayout
-}
-
-export let ViewLayoutsPanel = ({ schema }) => (view) => {
-    let ViewLayoutsPanelComponent = (props) => {
-        useRenderCount("ViewLayoutsPanel");
-
-        let { config } = view;
-
-        let formActions = useContext(ContextReduxFormActions); // TODO: make these generic
-        let formSelectors = useContext(ContextReduxFormSelectors);
-
-        let models = useContext(ContextRelativeModels);
-        let title = useInterpolatedString(models, config.title);
-
-        let simulationInfoPromise = useContext(ContextSimulationInfoPromise);
-
-        let store = useStore();
-
-        let getModels = () => {
-            return models.getModels(store.getState());
-        }
-
-        let saveToServer = (models) => {
-            simulationInfoPromise.then((simulationInfo) => {
-                simulationInfo.models = models;
-                fetch("/save-simulation", {
-                    method: 'POST',
-                    headers: {
-                        'Content-Type': 'application/json'
-                    },
-                    body: JSON.stringify(simulationInfo)
-                }).then(resp => {
-                    // TODO: error handling
-                    console.log("resp", resp);
-                })
-            })
-        }
-
-        if(!config) {
-            throw new Error("view missing config: " + view.name);
-        }
-
-        let basic = config.basic || [];
-        let advanced = config.advanced || [];
-
-        let dependencies = [...basic, ...advanced].map(layoutConfig => {
-            let ele = elementForLayoutName(layoutConfig.layout);
-            return ele.getDependencies(layoutConfig);
-        }).flat().map(dependencyString => new Dependency(dependencyString));
-
-        let hookedDependencyGroup = new HookedDependencyGroup({ schemaModels: schema.models, models, dependencies });
-
-        let hookedDependencies = dependencies.map(hookedDependencyGroup.getHookedDependency);
-
-        let formController = new FormController({ formActions, formSelectors, hookedDependencies });
-
-        let submit = () => {
-            let models = getModels();
-            formController.submitChanges();
-            saveToServer(models);
-        }
-
-        let mapLayoutToElement = (layoutConfig, idx) => {
-            let ele = elementForLayoutName(layoutConfig.layout);
-            let LayoutElement = ele.element;
-            return <LayoutElement key={idx} config={layoutConfig}></LayoutElement>;
-        }
-
-        let mainChildren = (!!basic) ? basic.map(mapLayoutToElement) : undefined;
-        let modalChildren = (!!advanced) ? advanced.map(mapLayoutToElement) : undefined;
-
-        let formProps = {
-            submit: submit,
-            cancel: formController.cancelChanges,
-            showButtons: formController.isFormStateDirty(),
-            formValid: formController.isFormStateValid(),
-            mainChildren,
-            modalChildren,
-            title: title || view.name,
-            id: view.name
-        }
-
-        return (
-            <ContextRelativeFormDependencies.Provider value={hookedDependencies}>
-                <ContextRelativeHookedDependencyGroup.Provider value={hookedDependencyGroup}>
-                    <ContextRelativeFormController.Provider value={formController}>
-                        <EditorPanel {...formProps}>
-                        </EditorPanel>
-                    </ContextRelativeFormController.Provider>
-                </ContextRelativeHookedDependencyGroup.Provider>
-            </ContextRelativeFormDependencies.Provider>
-        )
-    }
-    return ViewLayoutsPanelComponent;
 }
