+++ react/src/oldcomponents/panel.scss
