+++ react/src/views/missing.js
@@ -1,0 +1,9 @@
+export class MissingLayout extends View {
+    getFormDependencies = () => {
+        return [];
+    }
+
+    component = (props) => {
+        return <>Missing layout!</>;
+    }
+}
