+++ react/src/views/utility.js
@@ -1,0 +1,12 @@
+export function SpacedLayout(subView) {
+    return new class extends subView {
+        component = (props) => {
+            let ChildComponent = super.component;
+            return (
+                <div className="sr-form-layout">
+                    <ChildComponent {...props} />
+                </div>
+            )
+        }
+    }
+}
